Put helper functions into lib/util.js
