+++ lib/index.js
@@ -8,6 +8,9 @@ var parseTime = require('parse-messy-time')
 var strftime = require('strftime')
 var through2 = require('through2')
 
+var util = require('./util')
+var getDate = util.getDate
+
 function Clocker (options) {
   this.dir = options.dir
 
@@ -498,27 +501,4 @@ function getKey (x) {
 
 function toStamp (s) {
   return Math.floor(new Date(s.split('!')[1]).valueOf() / 1000)
-}
-
-function isUnixTimestamp (expr) {
-  return /^1[0-9]{9}$/.test(expr)
-}
-
-function getDate (expr) {
-  var d
-  var timestamp = Date.parse(expr)
-
-  if (expr === null || expr === undefined) {
-    d = new Date()
-  } else if (typeof expr === 'object' && expr instanceof Date) {
-    return expr
-  } else if (isUnixTimestamp(expr)) {
-    d = new Date(expr * 1000)
-  } else if (isNaN(timestamp)) {
-    d = parseTime(expr)
-  } else {
-    d = new Date(timestamp)
-  }
-
-  return d
 }
